fix: 통계 오버플로우 해결 by 우영

+++ back/app/src/main/java/com/sloway/app/payment/pay/repository/PayRepositoryImpl.java
@@ -54,7 +54,7 @@ public class PayRepositoryImpl implements PayRepositoryCustom {
                 .select(
                         qPayEntity.method,
                         qPayEntity.count(),
-                        qPayEntity.finalAmt.sum()
+                        qPayEntity.finalAmt.sum().longValue()
                 )
                 .from(qPayEntity)
                 .where(
